*Edit from last time: I fixed the thing that made the oranges all in a diagonal. Bottom problem is still a thing
So, we have a problem. Run this and move to a different section of the map, and you'll see that the oranges behave really oddly (they move with the map). Idk how you did the movement between sections, so I don't know how to fix this. Can you comment what you did with the movement so I can fix this? Thx.